refactor prediction endpoint
